Refactor: UserPrincipal위치 변경
