+++ back/src/main/java/com/eof/back/domain/gamesession/controller/GameSessionController.java
@@ -5,7 +5,7 @@ import com.eof.back.domain.gamesession.dto.GameSessionCreateResponse;
 import com.eof.back.domain.gamesession.dto.GameSessionJoinResponse;
 import com.eof.back.domain.gamesession.dto.GameSessionListResponse;
 import com.eof.back.domain.gamesession.service.GameSessionService;
-import com.eof.back.domain.user.dto.UserPrincipal;
+import com.eof.back.global.jwt.UserPrincipal;
 import com.eof.back.domain.user.entity.User;
 import com.eof.back.global.response.CommonResponse;
 import com.eof.back.global.response.Response;
