+++ back/src/main/java/com/eof/back/global/jwt/JwtAuthenticationFilter.java
@@ -1,6 +1,6 @@
 package com.eof.back.global.jwt;
 
-import com.eof.back.domain.user.dto.UserPrincipal;
+import com.eof.back.global.jwt.UserPrincipal;
 import com.eof.back.global.exception.exceptions.AuthException;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
