+++ back/src/test/java/com/eof/back/global/jwt/JwtAuthenticationFilterTest.java
@@ -1,6 +1,6 @@
 package com.eof.back.global.jwt;
 
-import com.eof.back.domain.user.dto.UserPrincipal;
+import com.eof.back.global.jwt.UserPrincipal;
 import com.eof.back.global.exception.errorCode.AuthErrorCode;
 import com.eof.back.global.exception.exceptions.AuthException;
 import io.jsonwebtoken.Claims;
